Fix type checking of pop/push to stack
InstructionSequence now checks what is popped from the stack matches what was previously pushed. And resolves to an eventual final pushed type.

+++ test/type_check_test.exs
@@ -105,7 +105,7 @@ defmodule TypeCheckTest do
     use Orb
 
     defw example(), a: F32 do
-      a = I32.when?(1, do: 1, else: 2)
+      a = I32.when?(i32(1), do: 1, else: 2)
     end
   end
